Shop/requirements.txt
Django==4.2.7  

gunicorn==21.2.0

djangorestframework>=3.12

djangorestframework-simplejwt>=5.2